+++ CHANGES.md
@@ -60,6 +60,7 @@ UI:
 -   Made files that are dropped get uploaded, if this is specified by the user
 -   Bugfix: Read raw value from CSV files
 -   Add a flag `placeholderWorkflowsOn` to hide UI components that are not complete yet
+-   Always set dataset publishing state to "published" no matter the approval flow is on / off
 
 Storage: